Make onnx-modifier installable and publish to PyPI (#121)
