+++ .gitignore
@@ -14,4 +14,6 @@ gym/
 local_tmp/
 *ppt
 *pptx
-local_test.py
+*.local.*
+README_pypi.md
+
