+++ README.md
@@ -32,7 +32,17 @@ Hope it helps!
 
 We have three methods to launch `onnx-modifier` now.
 
-## launch from command line
+## launch from installed wheel
+Install the pre-built wheel
+```bash
+pip install onnx-modifier
+```
+Then run
+```bash
+onnx-modifier
+```
+
+## launch from source code
 Clone the repo and install the required Python packages by
 
 ```bash
@@ -45,7 +55,7 @@ pip install -r requirements.txt
 Then run
 
 ```bash
-python app.py
+python entry.py
 ```
 
 Click the url in the output info generated by flask (defaults to `http://127.0.0.1:5000/`), then `onnx-modifier` will be launched in the web browser.
@@ -60,37 +70,6 @@ Click the url in the output info generated by flask (defaults to `http://127.0.0
 > I recorded how I made the the executable file in `app_desktop.py`. The executable file for other platforms are left for future work.
 
 </details>
-
-
-
-## launch from a docker container
-<details>
-  <summary>Click to expand</summary>
-
-We create a docker container like this:
-
-```bash
-git clone git@github.com:ZhangGe6/onnx-modifier.git
-cd onnx-modifier
-docker build --file Dockerfile . -t onnx-modifier
-```
-
-After building the container, we run onnx-modifier by mapping docker port and a local folder `modified_onnx`
-
-```bash
-mkdir -p modified_onnx
-docker run -d -t \
-  --name onnx-modifier \
-  -u $(id -u ${USER}):$(id -g ${USER}) \
-  -v $(pwd)/modified_onnx:/modified_onnx \
-  -p 5000:5000 \
-  onnx-modifier
-```
-
-Then we have access to onnx-modifer from URL <http://127.0.0.1:5000>. The modified ONNX models are expected to be found inside the local folder `modified_onnx`.
-</details>
-
-
 
 Click `Open Model...` to upload the ONNX model to edit. The model will be parsed and shown on the page.
