+++ onnx_modifier/static/view-sidebar.js
