+++ onnx_modifier/templates/index.html
@@ -441,7 +441,6 @@ warning_msg::before {
     <button id="add-node" class="graph-op-button-addNode">Add node</button>
     <select id="add-node-dropdown" class="graph-op-add-node-dropdown">
     </select>
-	<button id="load-model" class="graph-button-load">LoadModel</button>
 
     <button id="back-button" class="toolbar-back-button" title="Back">
         &#x276E;
